refactor: PostDetailPage 컴포넌트 확장자 변경
- MDX parser인 acorn이 Typescript와 호환되지 않기에 확장자를 tsx에서 jsx로 변경
